requirements.txt
fastapi
uvicorn 
httpx 
pydantic 
python-multipart  
requests

pandas
python-jose
pwdlib[argon2]